Estimation of the optimal subsample size for bagged CV bandwidth for Parzen-Rosenblatt estimator

Estimation of the optimal subsample size for bagged CV bandwidth for Parzen-Rosenblatt estimator

Usage

mopt(x, N, r = 1000, s = 100, ncores = parallel::detectCores())

Arguments

x

Vector. Sample.

N

Positive integer. Number of subsamples for the bagged bandwidth.

r

Positive integer. Size of the subsamples.

s

Positive integer. Number of subsamples.

ncores

Positive integer. Number of cores with which to parallelize the computations.

Value

Estimate of the optimal subsample size.

Details

Estimates the optimal size of the subsamples for the bagged CV bandwidth selector for the Parzen-Rosenblatt estimator.

Examples

set.seed(1)
x <- rt(10^5, 5)
mopt(x, 500, 500, 10, 2)
#> [1] 271